refactor(handlers): 重命名 mcp.handler 为 mcp-manage.handler
改动说明

- 为什么改：mcp.handler.ts 文件名无法准确反映其职责。该文件负责 MCP
服务的管理操作（添加、删除、查询服务），而非通用的 MCP 处理。为提高代码可读性和语义清晰度，将其重命名为
mcp-manage.handler.ts。
- 改了什么：
  - 重命名主文件：apps/backend/handlers/mcp.handler.ts → mcp-manage.handler.ts
  - 重命名测试文件：apps/backend/handlers/__tests__/mcp.handler.test.ts → mcp-manage.handler.test.ts
  - 更新导出路径：handlers/index.ts 中的导入路径从 ./mcp.handler.js 改为 ./mcp-manage.handler.js
  - 更新测试导入：测试文件中的导入路径从 ../mcp.handler.js 改为 ../mcp-manage.handler.js
- 影响范围：
  - 无破坏性变更：MCPHandler 类的公共 API 完全不变
  - 内部导入路径已更新：通过 @handlers/index.js 的引用者无需修改
  - 所有单元测试（567个）通过验证
- 验证方式：
  - ✅ TypeScript 类型检查通过
  - ✅ Biome Lint 检查通过
  - ✅ 所有单元测试通过（包含 MCPHandler 的 32 个测试用例）
  - ✅ 项目构建成功
