+++ apps/backend/handlers/index.ts
@@ -3,7 +3,7 @@ export * from "./CozeApiHandler.js";
 export * from "./HeartbeatHandler.js";
 export { EndpointHandler } from "./endpoint.handler.js";
 export * from "./MCPRouteHandler.js";
-export { MCPHandler } from "./mcp.handler.js";
+export { MCPHandler } from "./mcp-manage.handler.js";
 export * from "./RealtimeNotificationHandler.js";
 export * from "./ServiceApiHandler.js";
 export * from "./StaticFileHandler.js";
