+++ apps/backend/handlers/mcp-manage.handler.ts
